Remove migrations folder (this is a local requirement) add to .gitignore.
Improve the order of the installation instructions.

Installation instructions have been improved to follow the series of
steps to get up and running as quickly as possible.

Remove script auto-generated .json files from the repo

The files be auto-updated by any dev that downloads it to their computer
and runs build_scripts. May need a better long-term solution for
non-auto generated scripts in future.