Refactoring of the wooey_celery task views JSON
This is a restructuring of the celery task views JSON to return
a global queue, user queue and user results set. Each data set is
available as an individual JSON (maybe unneccessary) and also as
a combined all_queues_json for the timer update.

The navbar header now auto-refreshes the status of the queues
as jobs are executed.