Add support for disabling registration url link in template
A bit more complicated than expected. In order to test the state
of the setting we need to pull it into the template context, which
means having an assignment tag, e.g. get_wooey_setting <name> as <variable>

Django has a simple helper function for simple tag and assignment tags,
but not a combination. An implementation of this has been added
from: https://gist.github.com/natevw/f14812604be62c073461

It's possible to implement this without the helper but it's nicer to
have a simple wrapper as we may need it elsewhere.

With this we can now do both:

    {% get_wooey_setting WOOEY_REGISTER_URL %}
    {% get_wooey_setting WOOEY_REGISTER_URL as register_url %}

...with the former outputting the URL, the latter assigning to
variable. This is similar in practise to the Django url tag.

Checking on license for the include.