Job UUIDs for private URLs and public access
Implements a per job UUID that creates secure folders for jobs and
outputs (unguessable) and allows for sharing of results via a
unique ID with public access.

Sharing and permalinks are implemented using this approach.

Previous jobs will be given a UUID for access but their folders
and files will remain in place.