Evaluate Scripts non-lazy for group name: use default locale
Evaluate Scripts non-lazy in the wooey_settings file to use the
default system locale for this. Note that this will be overridden
and ignored if this value is set specifically by the user.

Translation has been removed from Wooey! as it's a brand-name.

The site tag remains translated with lazy evaluation, so it can
automatically match the visitors browser settings.